Throw specialised exceptions from store methods
I created checked exception types (KitNotFoundException,
DisguiseNotFoundException) which are intended to be thrown from the
KitStore#get and DisguiseStore#get methods when the kit or disguise
doesn't exist.

I felt having a checked exception here is better than an unchecked one
because the caller should catch this error, I don't want this error to
be fatal.