Rewrite the class SuperSmashKit as an interface
I know that I will be using this plugin to dynamically create
SuperSmashKits later, but this led me to immediately design
SuperSmashKit as a class that can be instantiated.

I realise that actually it's much more extensible to let SuperSmashKit
be an interface that other code can implement (allowing for easier
usages of my API in Java itself) and I can also continue dynamic
creation simply through

    new SuperSmashKit() {
      public Abilities getAbilities() {
        ...
      }
      ...
    }

I think it's much more sensible that SuperSmashKit be an interface.